14-06-2023 am2

+++ app/Http/Livewire/Admin/Mahjoza.php
@@ -17,6 +17,7 @@ class Mahjoza extends Component
   public $ShowDo=false;
   public $TajNo=0;
   public $BankRadio='wahda';
+  public $DelRadio='WithDel';
 
 
   protected $listeners = ['show'];
@@ -26,7 +27,8 @@ class Mahjoza extends Component
     $this->Show=$show;
   }
   public function Take(){
-   MahjozaModel::on(Auth()->user()->company)->truncate();
+    if ($this->DelRadio=='WithDel')
+      MahjozaModel::on(Auth()->user()->company)->truncate();
     $this->filename='c:\Excel\Mahjoza\\'.Auth()->user()->company.'.xlsx';
 
     $this->ShowDo=true;
